wip: separate commands
