+++ native/c/commands/core.hpp
@@ -1,0 +1,18 @@
+/**
+ * This program and the accompanying materials are made available under the terms of the
+ * Eclipse Public License v2.0 which accompanies this distribution, and is available at
+ * https://www.eclipse.org/legal/epl-v20.html
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Copyright Contributors to the Zowe Project.
+ *
+ */
+
+#include "../parser.hpp"
+
+namespace core
+{
+parser::Command &setup_root_command(int argc, char *argv[]);
+int execute_command(int argc, char *argv[]);
+} // namespace core
