+++ native/c/commands/tool.cpp
@@ -1,0 +1,459 @@
+/**
+ * This program and the accompanying materials are made available under the terms of the
+ * Eclipse Public License v2.0 which accompanies this distribution, and is available at
+ * https://www.eclipse.org/legal/epl-v20.html
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Copyright Contributors to the Zowe Project.
+ *
+ */
+
+#include "tool.hpp"
+#include "common_args.hpp"
+#include "../zcn.hpp"
+#include "../zds.hpp"
+#include "../zut.hpp"
+#include "../zuttype.h"
+#include <unistd.h>
+
+using namespace parser;
+using namespace std;
+using namespace commands::common;
+
+namespace tool
+{
+
+int handle_tool_convert_dsect(const ParseResult &result)
+{
+  int rc = 0;
+  ZCN zcn = {0};
+  unsigned int code = 0;
+  string resp;
+
+  string adata_dsn = result.get_value<std::string>("adata-dsn", "");
+  string chdr_dsn = result.get_value<std::string>("chdr-dsn", "");
+  string sysprint = result.get_value<std::string>("sysprint", "");
+  string sysout = result.get_value<std::string>("sysout", "");
+
+  const char *user = getlogin();
+  string struser(user);
+  transform(struser.begin(), struser.end(), struser.begin(), ::tolower);
+
+  if (sysprint.empty())
+    sysprint = "/tmp/" + struser + "_sysprint.txt";
+  if (sysout.empty())
+    sysout = "/tmp/" + struser + "_sysout.txt";
+
+  cout << adata_dsn << " " << chdr_dsn << " " << sysprint << " " << sysout << endl;
+
+  vector<string> dds;
+  dds.reserve(4);
+  dds.push_back("alloc fi(sysprint) path('" + sysprint + "') pathopts(owronly,ocreat,otrunc) pathmode(sirusr,siwusr,sirgrp) filedata(text) msg(2)");
+  dds.push_back("alloc fi(sysout) path('" + sysout + "') pathopts(owronly,ocreat,otrunc) pathmode(sirusr,siwusr,sirgrp) filedata(text) msg(2)");
+  dds.push_back("alloc fi(sysadata) da('" + adata_dsn + "') shr msg(2)");
+  dds.push_back("alloc fi(edcdsect) da('" + chdr_dsn + "') shr msg(2)");
+
+  rc = zut_loop_dynalloc(dds);
+  if (0 != rc)
+  {
+    return RTNCD_FAILURE;
+  }
+
+  rc = zut_convert_dsect();
+  if (0 != rc)
+  {
+    cerr << "Error: convert failed with rc: '" << rc << "'" << endl;
+    cout << "  See '" << sysprint << "' and '" << sysout << "' for more details" << endl;
+    return RTNCD_FAILURE;
+  }
+
+  cout << "DSECT converted to '" << chdr_dsn << "'" << endl;
+  cout << "Copy it via `cp \"//'" + chdr_dsn + "'\" <member>.h`" << endl;
+
+  // Free dynalloc dds
+  zut_free_dynalloc_dds(dds);
+
+  return rc;
+}
+
+int handle_tool_dynalloc(const ParseResult &result)
+{
+  int rc = 0;
+  unsigned int code = 0;
+  string resp;
+
+  string parm = result.get_value<std::string>("parm", "");
+
+  rc = zut_bpxwdyn(parm, &code, resp);
+  if (0 != rc)
+  {
+    cerr << "Error: bpxwdyn with parm '" << parm << "' rc: '" << rc << "'" << endl;
+    cerr << "  Details: " << resp << endl;
+    return RTNCD_FAILURE;
+  }
+
+  cout << resp << endl;
+
+  return rc;
+}
+
+int handle_tool_display_symbol(const ParseResult &result)
+{
+  int rc = 0;
+  string symbol = result.get_value<std::string>("symbol", "");
+  transform(symbol.begin(), symbol.end(), symbol.begin(), ::toupper);
+  symbol = "&" + symbol;
+  string value;
+  rc = zut_substitute_symbol(symbol, value);
+  if (0 != rc)
+  {
+    cerr << "Error: asasymbf with parm '" << symbol << "' rc: '" << rc << "'" << endl;
+    return RTNCD_FAILURE;
+  }
+  cout << value << endl;
+
+  return RTNCD_SUCCESS;
+}
+
+int handle_tool_search(const ParseResult &result)
+{
+  int rc = 0;
+
+  string pattern = result.get_value<std::string>("string", "");
+  string warn = result.get_value<std::string>("warn", "");
+  long long max_entries = result.get_value<long long>("max-entries", 0);
+  string dsn = result.get_value<std::string>("dsn", "");
+
+  ZDS zds = {0};
+  bool results_truncated = false;
+
+  if (max_entries > 0)
+  {
+    zds.max_entries = max_entries;
+  }
+
+  // List members in a data set
+  vector<ZDSMem> members;
+  rc = zds_list_members(&zds, dsn, members);
+
+  // Note if results are truncated
+  if (RTNCD_WARNING == rc)
+  {
+    if (ZDS_RSNCD_MAXED_ENTRIES_REACHED == zds.diag.detail_rc)
+    {
+      results_truncated = true;
+    }
+  }
+
+  // Note failure if we can't list
+  if (RTNCD_SUCCESS != rc && RTNCD_WARNING != rc)
+  {
+    cerr << "Error: could not read data set: '" << dsn << "' rc: '" << rc << "'" << endl;
+    cerr << "  Details: " << zds.diag.e_msg << endl;
+    return RTNCD_FAILURE;
+  }
+
+  // Perform dynalloc
+  vector<string> dds;
+  dds.reserve(3);
+  dds.push_back("alloc dd(newdd) da('" + dsn + "') shr");
+  dds.push_back("alloc dd(outdd)");
+  dds.push_back("alloc dd(sysin)");
+
+  rc = zut_loop_dynalloc(dds);
+  if (0 != rc)
+  {
+    return RTNCD_FAILURE;
+  }
+
+  // Build super c selection criteria
+  string data = " SRCHFOR '" + pattern + "'\n";
+
+  for (vector<ZDSMem>::iterator it = members.begin(); it != members.end(); ++it)
+  {
+    data += " SELECT " + it->name + "\n";
+  }
+
+  // Write control statements
+  zds_write_to_dd(&zds, "sysin", data);
+  if (0 != rc)
+  {
+    cerr << "Error: could not write to dd: '" << "sysin" << "' rc: '" << rc << "'" << endl;
+    cerr << "  Details: " << zds.diag.e_msg << endl;
+    return RTNCD_FAILURE;
+  }
+
+  // Perform search
+  rc = zut_search("parms are unused for now but can be passed to super c, e.g. ANYC (any case)");
+  if (RTNCD_SUCCESS != rc &&
+      RTNCD_WARNING != rc &&
+      ZUT_RTNCD_SEARCH_SUCCESS != rc &&
+      ZUT_RTNCD_SEARCH_WARNING != rc)
+  {
+    cerr << "Error: could error invoking ISRSUPC rc: '" << rc << "'" << endl;
+  }
+
+  // Read output from super c
+  string output;
+  rc = zds_read_from_dd(&zds, "outdd", output);
+  if (0 != rc)
+  {
+    cerr << "Error: could not read from dd: '" << "outdd" << "' rc: '" << rc << "'" << endl;
+    cerr << "  Details: " << zds.diag.e_msg << endl;
+    return RTNCD_FAILURE;
+  }
+  cout << output << endl;
+
+  if (results_truncated)
+  {
+    if (warn == "true")
+    {
+      cerr << "Warning: results truncated" << endl;
+    }
+  }
+
+  // Free dynalloc dds
+  zut_free_dynalloc_dds(dds);
+
+  return RTNCD_SUCCESS;
+}
+
+int handle_tool_amblist(const ParseResult &result)
+{
+  int rc = 0;
+
+  string dsn = result.get_value<std::string>("dsn", "");
+  string statements = " " + result.get_value<std::string>("control-statements", "");
+
+  // Perform dynalloc
+  vector<string> dds;
+  dds.reserve(3);
+  dds.push_back("alloc dd(syslib) da('" + dsn + "') shr");
+  dds.push_back("alloc dd(sysprint) lrecl(80) recfm(f,b) blksize(80)");
+  dds.push_back("alloc dd(sysin) lrecl(80) recfm(f,b) blksize(80)");
+
+  rc = zut_loop_dynalloc(dds);
+  if (0 != rc)
+  {
+    return RTNCD_FAILURE;
+  }
+
+  transform(statements.begin(), statements.end(), statements.begin(), ::toupper);
+
+  // Write control statements
+  ZDS zds = {0};
+  zds_write_to_dd(&zds, "sysin", statements);
+  if (0 != rc)
+  {
+    cerr << "Error: could not write to dd: '" << "sysin" << "' rc: '" << rc << "'" << endl;
+    cerr << "  Details: " << zds.diag.e_msg << endl;
+    return RTNCD_FAILURE;
+  }
+
+  // Perform search
+  rc = zut_run("AMBLIST");
+  if (RTNCD_SUCCESS != rc)
+  {
+    cerr << "Error: could error invoking AMBLIST rc: '" << rc << "'" << endl;
+  }
+
+  // Read output from amblist
+  string output;
+  rc = zds_read_from_dd(&zds, "sysprint", output);
+  if (0 != rc)
+  {
+    cerr << "Error: could not read from dd: '" << "sysprint" << "' rc: '" << rc << "'" << endl;
+    cerr << "  Details: " << zds.diag.e_msg << endl;
+    return RTNCD_FAILURE;
+  }
+  cout << output << endl;
+
+  // Free dynalloc dds
+  zut_free_dynalloc_dds(dds);
+
+  return RTNCD_SUCCESS;
+}
+
+int handle_tool_run(const ParseResult &result)
+{
+  int rc = 0;
+  string program = result.get_value<std::string>("program", "");
+  string dynalloc_pre = result.get_value<std::string>("dynalloc-pre", "");
+  string dynalloc_post = result.get_value<std::string>("dynalloc-post", "");
+
+  // Allocate anything that was requested
+  if (dynalloc_pre.length() > 0)
+  {
+    vector<string> dds;
+
+    ifstream in(dynalloc_pre.c_str());
+    if (!in.is_open())
+    {
+      cerr << "Error: could not open '" << dynalloc_pre << "'" << endl;
+      return RTNCD_FAILURE;
+    }
+
+    string line;
+    while (getline(in, line))
+    {
+      dds.push_back(line);
+    }
+    in.close();
+
+    rc = zut_loop_dynalloc(dds);
+    if (0 != rc)
+    {
+      return RTNCD_FAILURE;
+    }
+  }
+
+  string indd = result.get_value<std::string>("in-dd", "");
+  if (indd.length() > 0)
+  {
+    string ddname = "DD:" + indd;
+    ofstream out(ddname.c_str());
+    if (!out.is_open())
+    {
+      cerr << "Error: could not open input '" << ddname << "'" << endl;
+      return RTNCD_FAILURE;
+    }
+
+    string input = result.get_value<std::string>("input", "");
+    if (result.has("input"))
+    {
+      out << input << endl;
+    }
+
+    out.close();
+  }
+
+  transform(program.begin(), program.end(), program.begin(), ::toupper);
+
+  rc = zut_run(program);
+
+  if (0 != rc)
+  {
+    cerr << "Error: program '" << program << "' ended with rc: '" << rc << "'" << endl;
+    rc = RTNCD_FAILURE;
+  }
+
+  string outdd = result.get_value<std::string>("out-dd", "");
+  if (outdd.length() > 0)
+  {
+    string ddname = "DD:" + outdd;
+    ifstream in(ddname.c_str());
+    if (!in.is_open())
+    {
+      cerr << "Error: could not open output '" << ddname << "'" << endl;
+      return RTNCD_FAILURE;
+    }
+
+    string line;
+    while (getline(in, line))
+    {
+      cout << line << endl;
+    }
+    in.close();
+  }
+
+  // Optionally free everything that was allocated
+  if (dynalloc_post.length() > 0)
+  {
+    vector<string> dds;
+
+    ifstream in(dynalloc_post.c_str());
+    if (!in.is_open())
+    {
+      cerr << "Error: could not open '" << dynalloc_post << "'" << endl;
+    }
+
+    string line;
+    while (getline(in, line))
+    {
+      dds.push_back(line);
+    }
+    in.close();
+
+    zut_loop_dynalloc(dds);
+  }
+
+  return rc;
+}
+
+void register_commands(parser::Command &root_command)
+{
+  auto tool_cmd = command_ptr(new Command("tool", "tool operations"));
+
+  // Convert DSECT subcommand
+  auto tool_convert_dsect_cmd = command_ptr(new Command("ccnedsct", "convert dsect to c struct"));
+  tool_convert_dsect_cmd->add_keyword_arg("adata-dsn",
+                                          make_aliases("--adata-dsn", "--ad"),
+                                          "input adata dsn", ArgType_Single, true);
+  tool_convert_dsect_cmd->add_keyword_arg("chdr-dsn",
+                                          make_aliases("--chdr-dsn", "--cd"),
+                                          "output chdr dsn", ArgType_Single, true);
+  tool_convert_dsect_cmd->add_keyword_arg("sysprint",
+                                          make_aliases("--sysprint", "--sp"),
+                                          "sysprint output", ArgType_Single, false);
+  tool_convert_dsect_cmd->add_keyword_arg("sysout",
+                                          make_aliases("--sysout", "--so"),
+                                          "sysout output", ArgType_Single, false);
+  tool_convert_dsect_cmd->set_handler(handle_tool_convert_dsect);
+  tool_cmd->add_command(tool_convert_dsect_cmd);
+
+  // Dynalloc subcommand
+  auto tool_dynalloc_cmd = command_ptr(new Command("bpxwdy2", "dynalloc command"));
+  tool_dynalloc_cmd->add_positional_arg("parm", "dynalloc parm string", ArgType_Single, true);
+  tool_dynalloc_cmd->set_handler(handle_tool_dynalloc);
+  tool_cmd->add_command(tool_dynalloc_cmd);
+
+  // Display symbol subcommand
+  auto tool_display_symbol_cmd = command_ptr(new Command("display-symbol", "display system symbol"));
+  tool_display_symbol_cmd->add_positional_arg("symbol", "symbol to display", ArgType_Single, true);
+  tool_display_symbol_cmd->set_handler(handle_tool_display_symbol);
+  tool_cmd->add_command(tool_display_symbol_cmd);
+
+  // Search subcommand
+  auto tool_search_cmd = command_ptr(new Command("search", "search members for string"));
+  tool_search_cmd->add_positional_arg(DSN);
+  tool_search_cmd->add_positional_arg("string", "string to search for", ArgType_Single, true);
+  tool_search_cmd->add_keyword_arg(MAX_ENTRIES);
+  tool_search_cmd->add_keyword_arg(WARN);
+  tool_search_cmd->set_handler(handle_tool_search);
+  tool_cmd->add_command(tool_search_cmd);
+
+  // Amblist subcommand
+  auto tool_amblist_cmd = command_ptr(new Command("amblist", "invoke amblist"));
+  tool_amblist_cmd->add_positional_arg("dsn", "data containing input load modules", ArgType_Single, true);
+  tool_amblist_cmd->add_keyword_arg("control-statements",
+                                    make_aliases("--control-statements", "--cs"),
+                                    "amblist control statements, e.g. listload output=map,member=testprog",
+                                    ArgType_Single, true);
+  tool_amblist_cmd->set_handler(handle_tool_amblist);
+  tool_cmd->add_command(tool_amblist_cmd);
+
+  // Run subcommand
+  auto tool_run_cmd = command_ptr(new Command("run", "run a program"));
+  tool_run_cmd->add_positional_arg("program", "name of program to run", ArgType_Single, true);
+  tool_run_cmd->add_keyword_arg("dynalloc-pre",
+                                make_aliases("--dynalloc-pre", "--dp"),
+                                "dynalloc pre run statements", ArgType_Single, false);
+  tool_run_cmd->add_keyword_arg("dynalloc-post",
+                                make_aliases("--dynalloc-post", "--dt"),
+                                "dynalloc post run statements", ArgType_Single, false);
+  tool_run_cmd->add_keyword_arg("in-dd",
+                                make_aliases("--in-dd", "--idd"),
+                                "input ddname", ArgType_Single, false);
+  tool_run_cmd->add_keyword_arg("input",
+                                make_aliases("--input", "--in"),
+                                "input", ArgType_Single, false);
+  tool_run_cmd->add_keyword_arg("out-dd",
+                                make_aliases("--out-dd", "--odd"),
+                                "output ddname", ArgType_Single, false);
+  tool_run_cmd->set_handler(handle_tool_run);
+  tool_cmd->add_command(tool_run_cmd);
+
+  root_command.add_command(tool_cmd);
+}
+} // namespace tool
